Check periodically for retransmittable packets. Bug #46.

+++ briar-android/src/org/briarproject/plugins/droidtooth/DroidtoothTransportConnection.java
@@ -64,11 +64,11 @@ class DroidtoothTransportConnection implements DuplexTransportConnection {
 			return plugin.getMaxFrameLength();
 		}
 
-		public long getMaxLatency() {
+		public int getMaxLatency() {
 			return plugin.getMaxLatency();
 		}
 
-		public long getMaxIdleTime() {
+		public int getMaxIdleTime() {
 			return plugin.getMaxIdleTime();
 		}
