+++ briar-api/src/org/briarproject/api/plugins/Plugin.java
@@ -15,10 +15,10 @@ public interface Plugin {
 	int getMaxFrameLength();
 
 	/** Returns the transport's maximum latency in milliseconds. */
-	long getMaxLatency();
+	int getMaxLatency();
 
 	/** Returns the transport's maximum idle time in milliseconds. */
-	long getMaxIdleTime();
+	int getMaxIdleTime();
 
 	/** Starts the plugin and returns true if it started successfully. */
 	boolean start() throws IOException;
@@ -39,7 +39,7 @@ public interface Plugin {
 	 * Returns the desired interval in milliseconds between calls to the
 	 * plugin's {@link #poll(Collection)} method.
 	 */
-	long getPollingInterval();
+	int getPollingInterval();
 
 	/**
 	 * Attempts to establish connections to contacts, passing any created
