+++ briar-core/src/org/briarproject/plugins/file/FileTransportWriter.java
@@ -31,11 +31,11 @@ class FileTransportWriter implements TransportConnectionWriter {
 		return plugin.getMaxFrameLength();
 	}
 
-	public long getMaxLatency() {
+	public int getMaxLatency() {
 		return plugin.getMaxLatency();
 	}
 
-	public long getMaxIdleTime() {
+	public int getMaxIdleTime() {
 		return plugin.getMaxIdleTime();
 	}
