+++ briar-core/src/org/briarproject/plugins/tcp/LanTcpPlugin.java
@@ -17,8 +17,8 @@ class LanTcpPlugin extends TcpPlugin {
 	static final TransportId ID = new TransportId("lan");
 
 	LanTcpPlugin(Executor ioExecutor, DuplexPluginCallback callback,
-			int maxFrameLength, long maxLatency, long maxIdleTime,
-			long pollingInterval) {
+			int maxFrameLength, int maxLatency, int maxIdleTime,
+			int pollingInterval) {
 		super(ioExecutor, callback, maxFrameLength, maxLatency, maxIdleTime,
 				pollingInterval);
 	}
